requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements_lock.txt requirements.in
#
grpcio==1.59.0
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.59.0
    # via -r requirements.in
protobuf==4.23.4
    # via
    #   -r requirements.in
    #   grpcio-tools

# The following packages are considered to be unsafe in a requirements file:
# setuptools
